Refactor DialogInput for TS
